a lot of cleanup, added pmp

+++ kinclude/csr.c
@@ -1,3 +1,4 @@
+#include "ktypes.h"
 #include "csr.h"
 
 #ifdef TIMECMP_IN_MEMORY
@@ -6,31 +7,31 @@
 #error "You set TIMECMP_IN_MEMORY but did not provide a memory addres in TIMECMP_MEM_ADDR!"
 #endif
 
-void write_mtimecmp(unsigned long long int mtimecmp)
+void write_mtimecmp(uint64 mtimecmp)
 {
-    unsigned int lo = mtimecmp & 0xffffffff;
-    unsigned int hi = mtimecmp >> 32;
+    uint32 lo = mtimecmp & 0xffffffff;
+    uint32 hi = mtimecmp >> 32;
 
-    __asm__ volatile(
-         "li t0, %0\n"
-         "sw %1, 0(t0)\n"
-         "sw %2, 4(t0)" ::
-         "i"(TIMECMP_MEM_ADDR), "r"(lo), "r"(hi)
+    __asm__ volatile (
+          "li t0, %0\n"
+          "sw %1, 0(t0)\n"
+          "sw %2, 4(t0)" ::
+          "i"(TIMECMP_MEM_ADDR), "r"(lo), "r"(hi)
     );
 }
 
 #else
 
-void write_mtimecmp(unsigned long long int mtimecmp)
+void write_mtimecmp(uint64 mtimecmp)
 {
-    unsigned int lower = mtimecmp & 0xffffffff;
-    unsigned int higher = mtimecmp >> 32;
+    uint32 lower = mtimecmp & 0xffffffff;
+    uint32 higher = mtimecmp >> 32;
 
-    __asm__ volatile(
-         "csrw %0, %2\n"
-         "csrw %1, %3" ::
-         "I"(CSR_MTIMECMP),"I"(CSR_MTIMECMPH),
-         "r"(lower), "r"(higher)
+    __asm__ volatile (
+          "csrw %0, %2\n"
+          "csrw %1, %3" ::
+          "I"(CSR_MTIMECMP),"I"(CSR_MTIMECMPH),
+          "r"(lower), "r"(higher)
     );
 }
